check the validity of each steps in the scripts to see what documents are lost in each step of the process from the download, to the text mining and finally to the merging with the rest of the variables

check the validity of each steps in the scripts to see what documents are lost in each step of the process from the download, to the text mining and finally to the merging with the rest of the variables

check_extract(path_urls = "../Betin_Collodel/2. Text mining IMF_data/datasets/urls docs/urls_Requests_Reviews_articleIV.RData",
  path_tf_idf = "../Betin_Collodel/2. Text mining IMF_data/datasets/tagged docs/tf_idf.RData",
  path_final_tf_idf = "../Betin_Collodel/2. Text mining IMF_data/datasets/tagged docs/tf_idf_database.RData")

Arguments

path_urls
the path to locate the dataframe of urls (output of 1. Consolidate_urls.R)

path_tf_idf:
path to locate the tf_idf matrix (output of 2. Run_text_mining.R)

path_final_tf_idf:
path to locate the final database containing the tf of the text mining and the rest of the variables merge after (output of 3. Clean_database.R)

Value

a list with a table given the proportion of the files that have not been properly computed by country the files that have an error and have not been downloaded and the files that have not been downloaded and merged