check if there are gaps greater than one year in the documents

check if there are gaps greater than one year in the documents

check_gaps_pdfs(
  general_path = "/Volumes/Elements/IMF documents/",
  only_large = TRUE
)

Arguments

general_path
path with one subdirectory per individual and for each of them a files subdirectory with pdfs

only_large
Logical. Consider only gaps greater than two years. Default is TRUE.

Value

a nested list: for every individual the number of gaps and a dataframe with the details.

Author

Manuel Betin, Umberto Collodel